Crear una carpeta general para las tablas de propiedades intelectuales.
