+++ src/app/pages/administrador/administrador.module.ts
@@ -1,16 +1,16 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
 
-import { AdministradorRoutingModule } from './administrador-routing.module';
+import {AdministradorRoutingModule} from './administrador-routing.module';
 import {FormsModule} from '@angular/forms';
 import {NgbCollapseModule, NgbModalModule, NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
 import {SweetAlert2Module} from '@sweetalert2/ngx-sweetalert2';
 import {ApplicantListingComponent} from './applicant-listing/applicant-listing.component';
-import { ApplicantFormComponent } from './applicant-form/applicant-form.component';
+import {ApplicantFormComponent} from './applicant-form/applicant-form.component';
 import {CoordinatorListingComponent} from './coordinator-listing/coordinator-listing.component';
-import { CoordinatorFormComponent } from './coordinator-form/coordinator-form.component';
+import {CoordinatorFormComponent} from './coordinator-form/coordinator-form.component';
 import {ApplicantService, CoordinatorService, InstitucionService} from './shared-services';
-import { AdminDashboardComponent } from './admin-dashboard/admin-dashboard.component';
+import {AdminDashboardComponent} from './admin-dashboard/admin-dashboard.component';
 import {SharedModule} from '../../template/shared/shared.module';
 import {DropdownMenusModule} from '../../template/widgets';
 import {TranslationModule} from '../../modules/i18n';
@@ -46,4 +46,5 @@ import {TranslateModule} from '@ngx-translate/core';
     InstitucionService
   ]
 })
-export class AdministradorModule { }
+export class AdministradorModule {
+}
