+++ src/app/template/layout/core/layout.service.ts
@@ -17,7 +17,7 @@ import { ActivatedRoute } from '@angular/router';
 const LAYOUT_CONFIG_LOCAL_STORAGE_KEY = `${environment.appVersion}-layoutConfig`;
 const BASE_LAYOUT_TYPE_LOCAL_STORAGE_KEY = `${environment.appVersion}-baseLayoutType`;
 const defaultBaseLayoutType: LayoutType = 'light-sidebar';
-const defaultLayoutConfig: ILayout = DarkSidebarConfig;
+const defaultLayoutConfig: ILayout = LightSidebarConfig;
 
 export function getEmptyHTMLAttributes(): HTMLAttributesType {
   return {
@@ -153,14 +153,19 @@ export class LayoutService {
   getLayoutByType(layoutType: LayoutType | undefined): ILayout {
     switch (layoutType) {
       case 'dark-sidebar':
+        console.log('LayoutType:', DarkSidebarConfig);
         return DarkSidebarConfig;
       case 'light-sidebar':
+        console.log('LayoutType:', LightSidebarConfig);
         return LightSidebarConfig;
       case 'dark-header':
+        console.log('LayoutType:', DarkHeaderConfig);
         return DarkHeaderConfig;
       case 'light-header':
+        console.log('LayoutType:', LightHeaderConfig);
         return LightHeaderConfig;
       default:
+        console.log('LayoutType:', defaultLayoutConfig);
         return defaultLayoutConfig;
     }
   }
