+++ src/app/template/layout/layout.component.html
@@ -60,7 +60,7 @@
 <!-- end:: Drawers -->
 
 <!-- end:: Engage -->
-<app-engages></app-engages>
+<!--<app-engages></app-engages>-->
 <!-- end:: Engage -->
 
 <!-- begin:: Modals -->
